+++ src/sessions/p2p_session.rs
@@ -157,43 +157,33 @@ impl P2PSession {
     /// The created session will use the default socket type (currently UDP).
     ///
     /// # Errors
-    /// - Will return a `InvalidRequest` if the number of players is higher than the allowed maximum (see `MAX_PLAYERS`).
-    /// - Will return a `InvalidRequest` if `input_size` is higher than the allowed maximum (see `MAX_INPUT_BYTES`).
     /// - Will return `SocketCreationFailed` if the socket could not be created.
     pub fn new(num_players: u32, input_size: usize, local_port: u16) -> Result<Self, GGRSError> {
         // udp nonblocking socket creation
         let addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::UNSPECIFIED), local_port); //TODO: IpV6?
         let socket =
             Box::new(UdpNonBlockingSocket::new(addr).map_err(|_| GGRSError::SocketCreationFailed)?);
-        Self::new_impl(num_players, input_size, socket)
+        Ok(Self::new_impl(num_players, input_size, socket))
     }
 
     /// Creates a new `P2PSession` for players who participate on the game input. After creating the session, add local and remote players,
     /// set input delay for local players and then start the session. The session will use the provided socket.
-    ///
-    /// # Errors
-    /// - Will return a `InvalidRequest` if the number of players is higher than the allowed maximum (see `MAX_PLAYERS`).
-    /// - Will return a `InvalidRequest` if `input_size` is higher than the allowed maximum (see `MAX_INPUT_BYTES`).
     pub fn new_with_socket(
         num_players: u32,
         input_size: usize,
         socket: impl NonBlockingSocket + 'static,
-    ) -> Result<Self, GGRSError> {
+    ) -> Self {
         Self::new_impl(num_players, input_size, Box::new(socket))
     }
 
-    fn new_impl(
-        num_players: u32,
-        input_size: usize,
-        socket: Box<dyn NonBlockingSocket>,
-    ) -> Result<Self, GGRSError> {
+    fn new_impl(num_players: u32, input_size: usize, socket: Box<dyn NonBlockingSocket>) -> Self {
         // local connection status
         let mut local_connect_status = Vec::new();
         for _ in 0..num_players {
             local_connect_status.push(ConnectionStatus::default());
         }
 
-        Ok(Self {
+        Self {
             state: SessionState::Initializing,
             num_players,
             input_size,
@@ -210,7 +200,7 @@ impl P2PSession {
             disconnect_frame: NULL_FRAME,
             players: HashMap::new(),
             event_queue: VecDeque::new(),
-        })
+        }
     }
 
     /// Must be called for each player in the session (e.g. in a 3 player session, must be called 3 times) before starting the session. Returns the player handle